Create nowplaying.lua
This version of the extension was designed to be used with the VLC Icecast Radio Directory. This extension will write to a file the artist name and write to another file the track title. This is just the first version and future versions will have more compatibility and will have a dialog box.